[#59] Split commands/index.ts, add ConnectionManager tests, expand activation smoke tests
Split monolithic commands/index.ts (1598 lines) into 7 focused modules:
shared.ts, queryCommands.ts, tableCommands.ts, connectionCommands.ts,
historyCommands.ts, schemaCommands.ts, exportCommands.ts.

Add 70 unit tests for ConnectionManager covering CRUD, lifecycle,
multi-DB driver caching, folder management, color/readonly inheritance.

Add 6 activation smoke tests: tree views, test API, output channel,
SQL completion provider, native module resilience, deactivate safety.
